totalCount优化

+++ dk-foundation-engine/src/main/java/com/dk/foundation/engine/baseentity/PageResult.java
@@ -44,7 +44,7 @@ public class PageResult<T> implements Serializable {
 		this.totalPage = (int)Math.ceil((double)totalCount/pageSize);
 	}
 
-	public long getTotalCount() {
+	public int getTotalCount() {
 		return totalCount;
 	}
